Try init model directly in test_initialization
